pruebas Unitarias

+++ EjemploUnitarias/src/main/java/pe/edu/upeu/ServicioAImpl.java
@@ -17,6 +17,6 @@ public class ServicioAImpl implements ServicioA {
         if (num2 == 0) {
             throw  new ArithmeticException("No se puede dividir por cero");
         }
-        return num1 / num2;
+        return (double) num1 / num2;
     }
 }
